Change from ClassId to EmbeddedId
Change approach from ClassId to EmbeddedId for BookPublisher associative
entity. This does remove unique constrain problem which is present
in solution/byAlanHay branch.
This change is done according danny.lesnik question/answer:
https://stackoverflow.com/questions/6405746/mapping-manytomany-with-composite-primary-key-and-annotation
Solution is OK, but last problem is that DAO thing is done in one case
in converter in FixConfig class. This does cause that resulting
associative endpoint does present Book and Publisher data, but not
their links.